requirements.txt
anyio==3.7.1
ariadne==0.20.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
exceptiongroup==1.1.2
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
oauthlib==3.2.2
python-decouple==3.8
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.31.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
